trabalho-integracao/api/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pymongo
neo4j
redis
